fix(lockfile): remove lockfile
Originally, I added a lockfile to the project thinking that I needed to
make sure that no two instances run at once. I was under the impression
that the underlying backup tool wouldn't handle that. Turns out that
this is not true. For example, restic has its own locking capabilities.
Adding our own on top of that seems silly. The abstraction is not in the
right spot.