chore(e2e): add e2e around restic exec
This requires me to change the way stdout and stderr are handled. It
used to be that they were merged into stderr. Now they're split.

Back in the day, I was thinking that I would save stdout for commands
that require output from the backend. That's a neat idea but I don't
think that it's worth mangling the standard stream over it.